todo funct v1.0.0

+++ functions/APIs/todos.js
@@ -2,6 +2,7 @@ const { db } = require("../util/admin");
 
 exports.getAllTodos = (request, response) => {
   db.collection("todos")
+    .where("username", "==", request.user.username)
     .orderBy("createdAt", "desc")
     .get()
     .then((data) => {
@@ -40,6 +41,7 @@ exports.postOneTodo = (request, response) => {
   }
 
   const newTodoItem = {
+    username: request.user.username,
     title: request.body.title,
     body: request.body.body,
     createdAt: new Date().toISOString(),
@@ -64,6 +66,9 @@ exports.deleteTodo = (request, response) => {
     .then((doc) => {
       if (!doc.exists) {
         return response.status(404).json({ error: "Todo not found" });
+      }
+      if (doc.data().username !== request.user.username) {
+        return response.status(403).json({ error: "UnAuthorized" });
       }
       return document.delete();
     })
@@ -91,5 +96,30 @@ exports.editTodo = (request, response) => {
       return response.status(500).json({
         error: err.code,
       });
+    });
+};
+
+exports.getOneTodo = (request, response) => {
+  const document = db.doc(`/todos/${request.params.todoId}`);
+  document
+    .get()
+    .then((doc) => {
+      if (doc.data().username !== request.user.username) {
+        return response.status(403).json({ error: "UnAuthorized" });
+      }
+      if (!doc.exists) {
+        return response.status(404).json({ error: "Todo not found" });
+      }
+      const todos = {
+        todoId: doc.id,
+        title: doc.data().title,
+        body: doc.data().body,
+        createdAt: doc.data().createdAt,
+      }
+      return response.json(todos);
+    })
+    .catch((err) => {
+      console.error(err);
+      return response.status(500).json({ error: err.code });
     });
 };
